+++ src/cmd/ksh93/sh/path.c
@@ -1252,7 +1252,7 @@ static_fn void exscript(Shell_t *shp, char *path, char *argv[], char *const *env
         err = errno;
     }
     if ((euserid = geteuid()) != shp->gd->userid) {
-        n = strlcpy(name + 9, fmtbase((long)getpid(), 10, 0), sizeof(name) - 9);
+        n = strlcpy(name + 9, fmtbase(getpid(), 10, 0), sizeof(name) - 9);
         if (n >= sizeof(name) - 9) abort();  // this can't happen
         // Create a suid open file with owner equal effective uid.
         n = sh_open(name, O_CREAT | O_TRUNC | O_WRONLY | O_CLOEXEC, S_ISUID | S_IXUSR);
